Added the update/merged controller
This is intended to be the final version (WIP as of the moment) of the controller, with the wiring completely overhauled and being run on one Arduino Mega. The code uses interrupts and tracks time passed in order to effectively take in 6 channels of 55 hz pwm and output 8 channels of 55 hz pwm.